Rewrite abstract.
This does some mayor rewriting. Each paragraph has a short sentence that
summarizes it/gives it a title.

I have also deleted all the things about old-fashioned databases. We can
use the space we have won to explain in more detail why datreant is so cool.

I generally introduced more links.

There is a start for a paragraph about MDSynthesis

In rewriting I tried to follow the comments an old adviser gave me. That
is I should test if a word/sentence does work. A word does work if the
meaning of the text changes drastically or becomes not understandable.
This resulted here in shortening a lot of sentences or scratching them
completely. I like this because it makes the text clearer for me and
easier to read at the same time.